Create UART.h and refactor Project
