+++ Project/Header/UART.c
@@ -39,11 +39,7 @@ void initUART2(uint32_t baud_rate) {
 	UART2->C2 |= ((UART_C2_TE_MASK) | (UART_C2_RE_MASK));
 }	
 
-void offRGB() {
-	/* Active low, so set bits to 1 to switch off */
-	PTB->PSOR = MASK(RED_LED) | MASK(GREEN_LED);
-	PTD->PSOR = MASK(BLUE_LED);
-}
+
 
 // UART2 Transmit Poll
 void UART2_Transmit_Poll(uint8_t data) {
@@ -58,42 +54,4 @@ uint8_t UART2_Receive_Poll(void) {
 	// while not empty can read the data
 	while(!(UART2->S1 & UART_S1_RDRF_MASK));
 	return (UART2->D);
-}
-
-/* GPIO Initialization */
-void InitGPIO(void){
-	
-	// Enable Clock to PORTB and PORTD
-	SIM->SCGC5 |= ((SIM_SCGC5_PORTB_MASK) | (SIM_SCGC5_PORTD_MASK));
-	// Configure MUX settings to make all 3 pins GPIO
-	// PORT_PCR_MUX_MASK = 0x700u. Need to clear it first 
-	PORTB->PCR[RED_LED] &= ~PORT_PCR_MUX_MASK;
-	// Set PCR[COLOR_LED] to 1 because GPIO (pg 184 datasheet)
-	PORTB->PCR[RED_LED] |= PORT_PCR_MUX(1);
-	PORTB->PCR[GREEN_LED] &= ~PORT_PCR_MUX_MASK;
-	PORTB->PCR[GREEN_LED] |= PORT_PCR_MUX(1);
-	PORTD->PCR[BLUE_LED] &= ~PORT_PCR_MUX_MASK;
-	PORTD->PCR[BLUE_LED] |= PORT_PCR_MUX(1);
-	// Set Data Direction Registers for PortB and PortD
-	PTB->PDDR |= (MASK(RED_LED) | MASK(GREEN_LED));
-	PTD->PDDR |= MASK(BLUE_LED);
-	
-	
-}
-
-void ledControl(color_t color) {
-	offRGB();
-	switch(color) {
-		case RED:
-			PTB->PCOR |= MASK(RED_LED);
-			break;
-		case GREEN:
-			PTB->PCOR |= MASK(GREEN_LED);
-			break;
-		case BLUE:
-			PTD->PCOR |= MASK(BLUE_LED);
-			break;
-		default:
-			offRGB();
-	}
 }
