+++ Project/myBluetoothLED.c
@@ -5,16 +5,20 @@
 #include "RTE_Components.h"
 #include  CMSIS_device_header
 #include "cmsis_os2.h"
-#include "BTtoLED.h"
+#include "UART.h"
 #include "GPIO.h"
- 
- void pinControl(int val) {
-	if(val) {
-		PTE->PSOR |= MASK(20);
-	} else {
-		PTE->PCOR |= MASK(20);
-	}	
- }
+#include "LED.h"
+
+#define PTE20 20
+
+/* GPIO Initialization */
+void InitGPIO() {
+  InitLEDGPIO();
+  
+  uint8_t pinsE[] = {PTE20}, numPinsE = 1;
+  InitGPIOAll(PORTE, pin, numPinsE);
+}
+
 /*----------------------------------------------------------------------------
  * Application main thread
  *---------------------------------------------------------------------------*/
@@ -27,19 +31,14 @@ void app_main (void *argument) {
 		// RX
 		rx_data = UART2_Receive_Poll();
 		if (rx_data == 0x05) {
-			ledControl(RED);
-			pinControl(1);
-			osDelay(500);
-			offRGB();
-			pinControl(0);
-			osDelay(500);
-			ledControl(RED);
-			pinControl(1);
-			osDelay(500);
-			offRGB();
-			pinControl(0);
-			osDelay(500);
-			
+      for (uint8_t i = 0; i < 2; ++i) {
+        ledControl(RED, LED_ON);
+        setPin(PORTE, PTE20, HIGH)
+        osDelay(500);
+        offRGB();
+        setPin(PORTE, PTE20, LOW);
+        osDelay(500);
+      }			
 		}	
 	}
 }
@@ -50,20 +49,6 @@ int main (void) {
   SystemCoreClockUpdate();
 	initUART2(BAUD_RATE);
 	InitGPIO();
-	
-	// Configure MUX settings to make all 3 pins GPIO
-	// PORT_PCR_MUX_MASK = 0x700u. Need to clear it first 
-	PORTE->PCR[20] &= ~PORT_PCR_MUX_MASK;
-	// Set PCR[COLOR_LED] to 1 because GPIO (pg 184 datasheet)
-	PORTE->PCR[20] |= PORT_PCR_MUX(1);
-	PTE->PDDR |= MASK(20);
-	PTE->PCOR |= MASK(20);
-	//pinControl(1);
-	//uint8_t pin[] = {20};
-	//InitGPIOAll(PORTE, pin, 1);
-	//PTE->PSOR |= MASK(20);
-	//PTE->PCOR |= MASK(20);
-  // ...
  
   osKernelInitialize();                 // Initialize CMSIS-RTOS
   osThreadNew(app_main, NULL, NULL);    // Create application main thread
